[SYCL] Address review comments.
