+++ sycl/include/CL/sycl/detail/cg_types.hpp
@@ -1,3 +1,11 @@
+//==---- cg_types.hpp - Auxiliary types required by command group class ----==//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
 #pragma once
 
 #include <CL/sycl/detail/host_profiling_info.hpp>
