+++ sycl/include/CL/sycl/detail/handler_proxy.hpp
@@ -1,3 +1,11 @@
+//==--------- handler_proxy.hpp - Proxy methods to call in handler ---------==//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
 #pragma once
 
 #include <CL/sycl/access/access.hpp>
