+++ sycl/include/CL/sycl/interop_handler.hpp
@@ -1,3 +1,11 @@
+//==------- interop_handler.hpp - Argument for codeplay_introp_task --------==//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
 #pragma once
 
 #include <CL/sycl/accessor.hpp>
