+++ sycl/source/detail/handler_proxy.cpp
@@ -1,3 +1,11 @@
+//==--------- handler_proxy.cpp - Proxy methods to call in handler ---------==//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
 #include <CL/sycl/detail/handler_proxy.hpp>
 
 #include <CL/sycl/handler.hpp>
